+++ socialvec/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = """Nir Lotan"""
 __email__ = 'nir.lotan@gmail.com'
-__version__ = '0.1.0'
+__version__ = '0.1.2'