+++ HISTORY.rst
@@ -6,3 +6,15 @@ History
 ------------------
 
 * First release on PyPI.
+
+0.1.1 (2022-09-29)
+------------------
+
+* Include config.yaml in the distribution.
+
+0.1.2 (2022-10-02)
+------------------
+
+* reaarange config.yaml
+* Support multiple versions of the SocialVec model
+* Fix bug when searching for similarity using username